measure text utility functions
